refactor: 소스코드 원복 및 리펙토링 1차 진행
